foi realizado o refactory nos serializadores assim como foi feita a adequação onde ele era utilizado.
closes #23
