+++ api.http
@@ -10,8 +10,8 @@ POST {{baseURL}}/api/states
 Content-Type: application/json
 
 {
-    "name": "Distrito Federal",
-    "abbreviation": "DF"
+    "name": "São Paulo",
+    "abbreviation": "SP"
 }
 
 ###
